initial store integration and refactor geoman initialization
